+++ euporie/libs/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Module for modified external libraries with necessary changes due to be upstreamed."""
-from . import commonmark_extensions
-
-__all__ = ["commonmark_extensions"]