+++ euporie/libs/commonmark_extensions/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Modified version of commonmark_extensions table module."""
-from .tables import Table  # type: ignore
-
-__all__ = ["Table"]