+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changlog
 Upcoming
 --------
 
+- Add support for LaTex equations in markdown using flatlatex
+- Markdown tables no longer expand to the full width of the display
 - Show menu item descriptions in the statusbar
 - Add keyboard shortcuts to the menu
 - Allow status bar to be hidden